backend/requirement.txt
fastapi
uvicorn
opencv-python
numpy
matplotlib
tensorflow
pydantic
shutil  
python-dotenv  
google-generativeai